pyproject.toml
[project]
name = "mango-metadata-from-tables"
version = "0.1.0"
description = "Use this package to process tabular files in which each row represents an iRODS data object and each column contains either an identifier or metadata to add to this data object. It supports plain text files and Excel files, which could be stored locally or in iRODS itself."
readme = "README.md"
authors = [
    { name = "Mariana Montes" },
    { name = "Jef Scheepers" }  
]
requires-python = ">=3.10"
dependencies = [
    "click==8.1.7",
    "decopatch==1.4.10",
    "defusedxml==0.7.1",
    "et-xmlfile==2.0.0",
    "exceptiongroup==1.3.0",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "makefun==1.16.0",
    "mango-mdschema==1.0.3",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "numpy==2.2.4",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pandas==2.2.3",
    "pluggy==1.6.0",
    "prettytable==3.16.0",
    "pygments==2.19.1",
    "pytest==8.4.1",
    "pytest-cases==3.9.1",
    "python-dateutil==2.9.0.post0",
    "python-irodsclient==3.2.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "rich==13.9.4",
    "six==1.17.0",
    "tomli==2.2.1",
    "typing-extensions==4.13.0",
    "tzdata==2025.2",
    "validators==0.35.0",
    "wcwidth==0.2.13",
]

[project.scripts]
mango-metadata-from-tables = "mango_metadata_from_tables.cli:mdtab"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cases>=3.9.1",
]